# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FDA Regulatory Intelligence Platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "description": "Return canonical metadata, versions, and known sections.",
        "operationId": "get_document_api_documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuidanceDocument"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/sections": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Sections",
        "description": "Return ordered parent-section navigation for a document.",
        "operationId": "get_sections_api_documents__document_id__sections_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SectionNavigationEntry"
                  },
                  "title": "Response Get Sections Api Documents  Document Id  Sections Get"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/passages/{chunk_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Passage",
        "description": "Resolve an exact cited passage within the requested document.",
        "operationId": "get_passage_api_documents__document_id__passages__chunk_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PassageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/artifacts/{artifact_id}/download": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Download Artifact",
        "description": "Stream an authorized preserved source artifact without exposing its object key.",
        "operationId": "download_artifact_api_documents__document_id__artifacts__artifact_id__download_get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Authentication required or token invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Caller is not authorized for this operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found or not visible to the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "A required service is unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentMetadata": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Stable internal FDA document identifier."
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "FDA guidance title."
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url",
            "description": "Canonical FDA source URL."
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus",
            "description": "Draft, Final, or Withdrawn."
          },
          "version_hash": {
            "type": "string",
            "title": "Version Hash",
            "description": "SHA-256 hash of the source document version."
          },
          "docket_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docket Number",
            "description": "FDA docket number when available."
          },
          "docket_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Docket Url",
            "description": "FDA docket URL when available."
          },
          "issuing_office": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuing Office",
            "description": "Issuing FDA center or office."
          },
          "center": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Center",
            "description": "FDA center from the catalog."
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic",
            "description": "Primary FDA guidance topic."
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics",
            "description": "All FDA guidance topics from the catalog."
          },
          "communication_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Communication Type",
            "description": "FDA communication type from the catalog."
          },
          "regulated_product": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulated Product",
            "description": "Regulated product category from the FDA catalog."
          },
          "published_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Date",
            "description": "Original publication date."
          },
          "issue_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Date",
            "description": "FDA issue date."
          },
          "fda_last_changed": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fda Last Changed",
            "description": "FDA catalog changed timestamp."
          },
          "comment_close": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment Close",
            "description": "Comment close date for draft guidance."
          },
          "open_comment": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open Comment",
            "description": "Whether the FDA catalog marks comments as open."
          },
          "lifecycle_state": {
            "$ref": "#/components/schemas/LifecycleState",
            "description": "Registry lifecycle state derived from catalog diffs.",
            "default": "active"
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated",
            "description": "Most recent FDA update date."
          },
          "cfr_references": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cfr References",
            "description": "CFR references identified for metadata filters."
          },
          "product_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Product Codes",
            "description": "FDA product codes identified for metadata filters."
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count",
            "description": "Number of pages in the source PDF."
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Retrieved At",
            "description": "Timestamp when the document was ingested."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "title",
          "source_url",
          "status",
          "version_hash"
        ],
        "title": "DocumentMetadata",
        "description": "Metadata describing a single FDA guidance document version."
      },
...
      "DocumentVersionResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "version_hash": {
            "type": "string",
            "title": "Version Hash"
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "lifecycle_state": {
            "$ref": "#/components/schemas/LifecycleState",
            "default": "active"
          },
          "fda_last_changed": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fda Last Changed"
          },
          "supersedes_version_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Version Hash"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "version_hash",
          "source_url",
          "status",
          "created_at"
        ],
        "title": "DocumentVersionResponse",
        "description": "Public version metadata without internal storage coordinates."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Envelope for all structured API error responses."
      },
...
      "GuidanceDocument": {
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/DocumentMetadata",
            "description": "Document-level metadata."
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/DocumentVersionResponse"
            },
            "type": "array",
            "title": "Versions",
            "description": "Known preserved source versions for this document."
          },
          "section_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Section Ids",
            "description": "Ordered normalized section identifiers in the document."
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/DocumentArtifact"
            },
            "type": "array",
            "title": "Artifacts",
            "description": "Authorized preserved-source descriptors for this document."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "metadata"
        ],
        "title": "GuidanceDocument",
        "description": "A parsed FDA guidance document with preserved structure."
      },
...
      "PassageResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Stable internal FDA document identifier."
          },
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id",
            "description": "Stable unique chunk identifier."
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id",
            "description": "Normalized section identifier."
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title",
            "description": "Human-readable section heading."
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number",
            "description": "Source PDF page number."
          },
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Exact passage text."
          },
          "source_url": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Source Url",
            "description": "FDA source URL."
          },
          "version_hash": {
            "type": "string",
            "title": "Version Hash",
            "description": "SHA-256 hash of the source document version."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "chunk_id",
          "text",
          "source_url",
          "version_hash"
        ],
        "title": "PassageResponse",
        "description": "A single cited passage resolved from a chunk id."
      },
...
      "SectionNavigationEntry": {
        "properties": {
          "section_id": {
            "type": "string",
            "title": "Section Id",
            "description": "Normalized section identifier."
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title",
            "description": "Human-readable section heading."
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number",
            "description": "First page number for this section."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "section_id"
        ],
        "title": "SectionNavigationEntry",
        "description": "A single section entry for document section-tree navigation."
      },
...
    }
  }
}